==> notebooks/01_process_raw_reviews.ipynb <==
import pandas as pd


df = pd.read_csv("../data/reviews_raw.csv", encoding="utf-8", header=None)
df.columns = ["title", "text", "rating", "category", "product_name", "product_link"]

print(f"No. of reviews: {df.shape[0]}")
df.head(5)


df_dedup = df.drop_duplicates().loc[:, ["title", "text", "rating"]]

print(f"No. of reviews: {df_dedup.shape[0]}")
df_dedup.head(5)


df_dedup.to_csv("../data/router_reviews.csv")
