==> notebooks/02_merge_reviews.ipynb <==
import pandas as pd


df1 = pd.read_csv("../data/Rahul's_CSV.csv")
df2 = pd.read_csv("../data/router_reviews.csv")
df3 = pd.read_csv("../data/7817_1.csv")
df4 = pd.read_csv("../data/Reviews.csv")


print(df1.columns)
print(df2.columns)

df1 = df1.loc[:, ["review_title", "description", "rating"]]
df1.columns = ["review_title", "review_text", "review_rating"]

df2 = df2.loc[:, ["title", "text", "rating"]]
df2.columns = ["review_title", "review_text", "review_rating"]


print(df3.columns)

df3 = df3.loc[:, ["reviews.title", "reviews.text", "reviews.rating"]]
df3.columns = ["review_title", "review_text", "review_rating"]
df3 = df3.dropna()
df3["review_rating"] = df3["review_rating"].astype(int)

df3


print(df4.columns)

df4 = df4.loc[:, ["Summary", "Text", "Score"]]
df4.columns = ["review_title", "review_text", "review_rating"]

df4


df_merged = pd.concat([df1, df2, df3, df4])

df_merged["review_type"] = df_merged["review_rating"].apply(
    lambda rating: "positive"
    if rating > 3
    else "neutral"
    if rating == 3
    else "negative"
)

df_merged["review_type"].value_counts()


df_merged


df = pd.concat(
    [
        df_merged.query("review_type == 'positive'").sample(n=10000),
        df_merged.query("review_type == 'neutral'").sample(n=10000),
        df_merged.query("review_type == 'negative'").sample(n=10000),
    ]
)


df.to_csv("../data/reviews_div.csv", index=False)
