==> notebooks/03_clean_data.ipynb <==
import pandas as pd
import contractions
import re

import nltk
from string import punctuation
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk import word_tokenize, pos_tag

from tqdm import tqdm

tqdm.pandas()


nltk.download("punkt")
nltk.download("wordnet")
nltk.download("stopwords")
nltk.download("averaged_perceptron_tagger")


df = pd.read_csv("../data/Merged/reviews_div.csv")
df["review_text"] = df["review_text"].astype(str)
df["review_title"] = df["review_title"].astype(str)

df


stop = set(stopwords.words("english"))
lemmatizer = WordNetLemmatizer()

pats = [re.compile("<.*?>")]


def clean_text(text: str, remove_stopwords_only: bool = False) -> str:
    if not remove_stopwords_only:
        text = str(contractions.fix(text))
        text = text.lower()
        text = re.sub(r"\s+", " ", text)
        for pat in pats:
            text = re.sub(pat, "", text)
        words = [
            lemmatizer.lemmatize(i, j[0].lower())
            if j[0].lower() in ["a", "n", "v"]
            else lemmatizer.lemmatize(i)
            for i, j in pos_tag(word_tokenize(text))
        ]
        text = " ".join(words)

    if remove_stopwords_only:
        text = text.translate(str.maketrans("", "", punctuation))
        words = [
            word
            for word in word_tokenize(text)
            if (word.lower() not in stop or word.lower() in {"go", "do", "no", "not"})
        ]

    return text


df["review_text_cleaned"] = df["review_text"].progress_map(clean_text)
df["review_title_cleaned"] = df["review_title"].progress_map(clean_text)


df.to_csv("../data/Merged/reviews_cleaned_nostop.csv", index=False)


def get_punct_count(text):
    return len(re.findall(rf"[{punctuation}]", text))


df["num_special_chars"] = df["review_text_cleaned"].apply(get_punct_count)


df["review_text_cleaned"] = df["review_text_cleaned"].progress_map(
    lambda text: clean_text(text, remove_stopwords_only=True)
)
df["review_title_cleaned"] = df["review_title_cleaned"].progress_map(
    lambda text: clean_text(text, remove_stopwords_only=True)
)


df["title_length"] = df["review_title_cleaned"].apply(lambda title: len(title.split()))
df["text_length"] = df["review_text_cleaned"].apply(lambda text: len(text.split()))


df = df.loc[
    :,
    [
        "review_title_cleaned",
        "review_text_cleaned",
        "title_length",
        "text_length",
        "num_special_chars",
        "review_type",
    ],
]
df.to_csv("../data/Merged/reviews_cleaned.csv", index=False)
