==> notebooks/05_vectorization_ml_models.ipynb <==
## Setup

import pandas as pd
from sklearn.naive_bayes import MultinomialNB
from xgboost import XGBClassifier
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import classification_report
from sklearn.preprocessing import LabelEncoder, OneHotEncoder
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.base import BaseEstimator
import numpy as np
from sklearn.tree import DecisionTreeClassifier
import gensim.downloader as gensim_api
from gensim.models import Word2Vec, KeyedVectors
from nltk.tokenize import word_tokenize

np.set_printoptions(precision=2, suppress=True)


df = pd.read_csv("../data/Merged/reviews_cleaned.csv")

df["review_title_cleaned"] = df["review_title_cleaned"].astype(str)
df["review_text_cleaned"] = df["review_text_cleaned"].astype(str)

df


def split_X_y(data: pd.DataFrame, encoder: LabelEncoder, fit: bool = False):
    features = data.drop("review_type", axis=1)
    labels = data.iloc[:, -1].to_numpy().reshape(-1, 1)

    if fit:
        target = encoder.fit_transform(labels)
    else:
        target = encoder.transform(labels)

    return features, np.array(target)


def build_model_pipe(
    vectorizer: CountVectorizer, model: BaseEstimator, include_extra: bool = True
) -> Pipeline:
    column_transformer = ColumnTransformer(
        [
            ("title_vectorizer", vectorizer, "review_title_cleaned"),
            ("text_vectorizer", vectorizer, "review_text_cleaned"),
        ],
        remainder="passthrough" if include_extra else "drop",
    )

    pipe = Pipeline([("vectorize", column_transformer), ("classify", model)])

    return pipe


encoder = LabelEncoder()

train_df = df.sample(frac=0.8)
X_train, y_train = split_X_y(train_df, encoder, fit=True)

test_df = df.drop(train_df.index)
X_test, y_test = split_X_y(test_df, encoder)


## Bag of Words

pipe_bow_mnb_clf = build_model_pipe(
    vectorizer=CountVectorizer(), model=MultinomialNB(), include_extra=False
)

pipe_bow_mnb_clf.fit(X_train, y_train)

y_pred = pipe_bow_mnb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


pipe_bow_xgb_clf = build_model_pipe(
    vectorizer=CountVectorizer(max_features=35000),
    model=XGBClassifier(),
    include_extra=False,
)

pipe_bow_xgb_clf.fit(X_train, y_train)

y_pred = pipe_bow_xgb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


import joblib
joblib.dump(pipe_bow_xgb_clf, "../data/models/bow_xgb_clf.jlib")

## Binary Bag of Words

pipe_bbow_mnb_clf = build_model_pipe(
    vectorizer=CountVectorizer(max_features=40000, binary=True),
    model=MultinomialNB(),
    include_extra=False,
)

pipe_bbow_mnb_clf.fit(X_train, y_train)

y_pred = pipe_bbow_mnb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


pipe_bbow_xgb_clf = build_model_pipe(
    vectorizer=CountVectorizer(max_features=45000, binary=True),
    model=XGBClassifier(),
    include_extra=False,
)

pipe_bbow_xgb_clf.fit(X_train, y_train)

y_pred = pipe_bbow_xgb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


## N-Grams

pipe_ngram_mnb_clf = build_model_pipe(
    vectorizer=CountVectorizer(ngram_range=(2, 3)), model=MultinomialNB()
)

pipe_ngram_mnb_clf.fit(X_train, y_train)

y_pred = pipe_ngram_mnb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


pipe_ngram_xgb_clf = build_model_pipe(
    vectorizer=CountVectorizer(ngram_range=(2, 5)), model=XGBClassifier()
)

pipe_ngram_xgb_clf.fit(X_train, y_train)

y_pred = pipe_ngram_xgb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


## TF-IDF

pipe_tfidf_mnb_clf = build_model_pipe(
    vectorizer=TfidfVectorizer(), model=MultinomialNB(), include_extra=False
)

pipe_tfidf_mnb_clf.fit(X_train, y_train)

y_pred = pipe_tfidf_mnb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


pipe_tfidf_xgb_clf = build_model_pipe(
    vectorizer=TfidfVectorizer(ngram_range=(2, 5), max_features=400000),
    model=XGBClassifier(),
)

pipe_tfidf_xgb_clf.fit(X_train, y_train)

y_pred = pipe_tfidf_xgb_clf.predict(X_test)
print(classification_report(y_test, y_pred))


## Word2Vec

w2v: KeyedVectors = gensim_api.load("glove-twitter-200")

def vectorize_sentence(text: str):
    words = word_tokenize(text)
    vectors = []
    for it in words:
        if it in w2v.key_to_index:
            vectors.append(w2v[it])
        else:
            vectors.append([0.0] * 200)
    return np.mean(vectors, axis=0).tolist()


X_train_vec = np.array(
    [vectorize_sentence(it) for it in X_train["review_text_cleaned"]]
)
X_test_vec = np.array([vectorize_sentence(it) for it in X_test["review_text_cleaned"]])


from sklearn.linear_model import RidgeClassifier

m = RidgeClassifier()
m.fit(X_train_vec, y_train)
y_pred = m.predict(X_test_vec)
print(classification_report(y_test, y_pred))


m = XGBClassifier()
m.fit(X_train_vec, y_train)
y_pred = m.predict(X_test_vec)
print(classification_report(y_test, y_pred))
