==> notebooks/06_dl_models.ipynb <==
%load_ext autotime 

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import tensorflow as tf
import keras
from keras import (
    activations,
    callbacks,
    preprocessing,
    layers,
    losses,
    metrics,
    models,
    optimizers,
)
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OneHotEncoder
from sklearn.metrics import (
    classification_report,
    RocCurveDisplay,
    ConfusionMatrixDisplay,
)

np.set_printoptions(precision=2, suppress=True)

import warnings

warnings.filterwarnings("ignore", "all")


data = pd.read_csv("../data/Merged/reviews_cleaned.csv")

MAX_FEATURES = 30_000
MAX_LEN = 200

def build_vectorizer() -> layers.TextVectorization:
    vec = layers.TextVectorization(
        max_tokens=MAX_FEATURES,
        output_sequence_length=MAX_LEN,
        pad_to_max_tokens=True,
        encoding="int",
    )

    return vec


vectorizer = build_vectorizer()
vectorizer.adapt(data["review_text_cleaned"])

def build_model(vectorizer: layers.TextVectorization) -> models.Model:
    text_input = keras.Input(shape=(1,), dtype=tf.string, name="text")
    X = vectorizer(text_input)
    X = layers.Embedding(MAX_FEATURES, 128)(X)

    X = layers.Bidirectional(layers.LSTM(64, return_sequences=True))(X)
    X = layers.Bidirectional(layers.LSTM(64))(X)
    outputs = layers.Dense(1, activation="sigmoid")(X)

    model = keras.Model(inputs=text_input, outputs=outputs)
    model.compile(
        loss=losses.categorical_crossentropy,
        optimizer=optimizers.RMSprop(learning_rate=1e-3),
        metrics=[metrics.AUC(name="auc"), metrics.CategoricalAccuracy(name="acc")],
    )

    return model


model = build_model(vectorizer)
model.summary()


X_train, X_test, y_train, y_test = train_test_split(
    data["review_text_cleaned"], data["review_type"], train_size=0.7
)

X_test, X_val, y_test, y_val = train_test_split(X_test, y_test, train_size=0.5)

enc = OneHotEncoder()
enc.fit(y_train.to_numpy().reshape(1, -1))

y_train = enc.transform(y_train)
y_test = enc.transform(y_test)
y_val = enc.transform(y_val)


callbacks = [
    callbacks.EarlyStopping(
        monitor="val_auc", mode="max", patience=10, restore_best_weights=True
    ),
    callbacks.ReduceLROnPlateau(monitor="val_auc", mode="max", factor=0.7, patience=4),
]

history = model.fit(
    X_train,
    y_train,
    validation_data=(X_val, y_val),
    epochs=60,
    batch_size=64,
    callbacks=callbacks,
)


def plot_hist(hist, metric='auc'):
    plt.plot(hist.history[metric])
    plt.plot(hist.history["val_" + metric])
    plt.title(f"model performance")
    plt.ylabel("area_under_curve")
    plt.xlabel("epoch")
    plt.legend(["train", "validation"], loc="upper left")
    plt.show()
    return
